requirements.txt
alembic==1.4.2
astroid==2.4.1
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.1
contextlib2==0.6.0
cryptography==2.8
datarobot==2.20.0
dominate==2.5.1
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jwt==0.6.1
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.5
pandas==1.0.4
psycopg2-binary==2.8.5
pycparser==2.20
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
requests==2.23.0
requests-toolbelt==0.9.1
rq==1.4.2
six==1.14.0
SQLAlchemy==1.3.15
toml==0.10.1
trafaret==1.2.0
typed-ast==1.4.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.0
wrapt==1.12.1
WTForms==2.2.1
